added testing, moved test file tinto test directory
